# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yunora AI Question Generation Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List all boards
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardList"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a board
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Created board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
  /boards/{id}:
    get:
      operationId: getBoard
      tags: [boards]
      summary: Get board by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
    patch:
      operationId: updateBoard
      tags: [boards]
      summary: Update board
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
    delete:
      operationId: deleteBoard
      tags: [boards]
      summary: Delete board
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── STANDARDS ────────────────────────────────────────────────────────────────
...
  /chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters
      security:
        - bearerAuth: []
      parameters:
        - name: subjectId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: syllabus
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterList"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/syllabus-categories:
    get:
      operationId: listChapterSyllabusCategories
      tags: [chapters]
      summary: List unique syllabus categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /chapters/{id}:
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update chapter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete chapter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── TOPICS ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Board:
      type: object
      required: [id, name, code]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        standardsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    BoardInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean, default: true }
    BoardUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    BoardList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Board"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── STANDARDS ─────────────────────────────────────────────────────────────
...
    Chapter:
      type: object
      required: [id, name, subjectId, orderIndex]
      properties:
        id: { type: integer }
        name: { type: string }
        orderIndex: { type: integer }
        subjectId: { type: integer }
        subjectName: { type: ["string", "null"] }
        syllabus: { type: ["string", "null"] }
        isActive: { type: boolean }
        topicsCount: { type: integer }
        questionsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    ChapterInput:
      type: object
      required: [name, subjectId, orderIndex]
      properties:
        name: { type: string }
        orderIndex: { type: integer }
        subjectId: { type: integer }
        syllabus: { type: ["string", "null"] }
        isActive: { type: boolean, default: true }
    ChapterUpdate:
      type: object
      properties:
        name: { type: string }
        orderIndex: { type: integer }
        syllabus: { type: ["string", "null"] }
        isActive: { type: boolean }
    ChapterList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── TOPICS ────────────────────────────────────────────────────────────────
...